refactor(docs): flatten src/lib into src/
Remove the lib/ subdirectory — for a small pipeline package,
the extra nesting adds indirection without meaningful grouping.

- Move all files from src/lib/ directly into src/
- Rename types.d.ts → types.ts (correct extension for module types)
- Inline capitalize() into generateMarkdownTable.ts and delete the file
