+++ packages/docs/src/extract.ts
