--- packages/docs/src/lib/capitalize.ts
@@ -1,1 +1,0 @@
-export const capitalize = (value: string) => value.charAt(0).toUpperCase() + value.slice(1);
